Added UI settings, withdrawal and preferences
The settings view now has links to a detailed message preference view,
privacy (not implemented yet), and a withdrawal link. The withdraw
function has a confirm popup, sends an http request, and displays a
failure message if no contact can be made with the backend. But at this
point there's nothing on the backend for withdrawing.