Handle withdrawal request on the backend
We do this by setting the flag on the user. Login then isn't allowed if
the user is withdrawn. Also, even if the auth token is valid (which it will
remain), it won't clear the SecureAPIController before: method if the
user is withdrawn. We can do this because the authtoken is exactly tied
to the user account on a 1:1 basis.

This also gives us the benefit that we can un-withdraw participants (if
necessary), and they'll be able to use the app without necessarily
having to log in again and get a new token.