Changes to text and presentation
Tidied up the category texts, because we'll be displaying them to
users. Attitude shifter is now "Thought for the day" and we will allow
it to be turned on and off, so I've taken the preferable switch off in
the seed file. Also, added 'being' to the front of all the 'goal' type
categories so that we don't do it in the view, which was silly.