Fixed nasty deletion bug
A tricky problem where deletion was crashing the app apparently at
random. What was happening was that I was splicing an array that I was
iterating over. Never, ever do this. Javascript doesn't warn you that
this isn't safe, but I was causing the index to be recalculated. I think
it's gone now...