Revert "Whitelist tel: links for the GP number"
This reverts commit db5c9f2756222fc51841947f027900fec4e4784d.

This seems to break Ionic/Angular UI routing with parameters. Don't know
why or how to fix it but there's precious little time, so for now I'm
removing the feature for allowing tel: links.

We don't really want people calling from the device anyway, they can
make a note of the number. If it's really wanted, I can try again and
see what's breaking it, but I think it's the whitelist.